requirements.txt
beautifulsoup4==4.5.1
codegen==1.0
coverage==4.2
Django==1.10.3
django-debug-toolbar==1.6
django-discover-runner==1.0
django-webtest==1.8.0
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
Markdown==2.6.7
mysqlclient==1.3.9
patterns==0.3
psycopg2==2.7.1
six==1.10.0
sqlparse==0.2.2
waitress==1.0.1
WebOb==1.6.2
WebTest==2.0.23
